% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dodge_x_labs}
\alias{dodge_x_labs}
\title{Wrapper function to dodge xAxis labels}
\usage{
dodge_x_labs(n.dodge = 2)
}
\arguments{
\item{n.dodge}{passed to \link[ggplot2:guide_axis]{ggplot2::guide_axis}}
}
\description{
Useful when you have overlapping labels on the xAxis.
}
