% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_bayesian_krig.R
\name{epi_bayesian_krig}
\alias{epi_bayesian_krig}
\title{Interpolate a grid using bayesian kriging (MCMC).}
\usage{
epi_bayesian_krig(epi, grid, mod, ...)
}
\arguments{
\item{epi}{a proper EpiVis data frame.}

\item{grid}{an \verb{n x 2} matrix of coordinates that one wish to predict at.}

\item{mod}{a string denoting the name of the stan file, but with exactly four \code{data} entries: nobs, npred, rate_obs, coord.}

\item{...}{additional parameters passed to rstan::sampling.}
}
\value{
fitted stan model.
}
\description{
Using coordinate grid and proper EpiVis dataframe to interpolate the grid with MCMC.
}
\details{
Note that this method could be significantly slow if too many points are to be calculated,
as calculating Covariance matrix is O(n^2). Also, the rate is multiplied by 10000 to avoid machine precision
errors.is a relatively low-level function allowing users to customize their model, chains,
and iterations. The output model and samples are saved locally for future use.
}
