% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_df.R
\name{epi_check}
\alias{epi_check}
\title{Check whether a data frame is a proper EpiVis data frame.}
\usage{
epi_check(data = NULL)
}
\arguments{
\item{data}{a data frame.}
}
\value{
Returns true if and only if the data frame is a proper EpiVis data frame.
}
\description{
Given a data frame, check if it is consistent with the requirements of a proper EpiVis data frame.
}
\details{
Note that a proper EpiVis data frame can be a proper superset of in terms of
variables (x, y, rate, map), but all these variables must have same length after
removing NA's.
}
