% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_df.R
\name{epi_converge}
\alias{epi_converge}
\title{Converge coordinates, rate and map into a proper EpiVis data frame.}
\usage{
epi_converge(
  x = NULL,
  y = NULL,
  rate = NULL,
  case = NULL,
  population = NULL,
  map = NULL
)
}
\arguments{
\item{x}{A vector of \code{n} coordinates of x}

\item{y}{A vector of \code{n} coordinates of y}

\item{rate}{A vector of \code{n} numerics denoting the rate of the cases (case/population)}

\item{case}{A vector of \code{n} numerics denoting the total number of cases}

\item{population}{A vector of \code{n} numerics denoting the number of population}

\item{map}{A string of name of the map of interest}
}
\value{
A proper EpiVis data frame that is compatible with EpiVis functions.
}
\description{
Requires all vectors to have the same length, map should be a just string (or a sequence of characters.)

Requires: 1. rate or 2. case and population; will converge to rate eventually.

x, y, rate, case, population are allowed to have NA's, but observations with any NA's will be thrown out.
}
