% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{epi_ggtheme}
\alias{epi_ggtheme}
\title{Epurated Ggplot theme (inspired from hrbrthemes)}
\usage{
epi_ggtheme(
  base_family = "sans",
  base_size = 10,
  plot_title_family = base_family,
  plot_title_size = 14,
  plot_title_face = "bold",
  plot_title_margin = 10,
  subtitle_family = base_family,
  subtitle_size = 9,
  subtitle_face = "italic",
  subtitle_margin = 15,
  strip_text_family = base_family,
  strip_text_size = 6,
  strip_text_face = "plain",
  caption_family = base_family,
  caption_size = 9,
  caption_face = "italic",
  caption_margin = 10,
  axis_text_size = base_size,
  axis_title_family = subtitle_family,
  axis_title_size = 10,
  axis_title_face = "plain",
  axis_title_just = "rt",
  plot_margin = margin(10, 10, 10, 10),
  grid_col = "#cccccc",
  grid = TRUE,
  axis_col = "#cccccc",
  axis = FALSE,
  ticks = FALSE
)
}
\arguments{
\item{base_family}{base font family}

\item{base_size}{base font size}

\item{plot_title_family, plot_title_size, plot_title_face, plot_title_margin}{plot title family, face, size and margins}

\item{subtitle_family, subtitle_size, subtitle_face, subtitle_margin}{plot subtitle family, face, size and margins}

\item{strip_text_family, strip_text_size, strip_text_face}{Facet label family, face, size}

\item{caption_family, caption_size, caption_size, caption_margin}{plot caption family, face, size and margin}

\item{axis_text_size}{font size of axis text}

\item{axis_title_family, axis_title_size, axis_title_face}{axis title font family, face and size}

\item{axis_title_just}{axis title font justification, one of ⁠\link{blmcrt}⁠}

\item{plot_margin}{plot margin (specify with ggplot2::margin())}

\item{grid_col, axis_col}{grid & axis colors; both default to ⁠#cccccc⁠}

\item{grid}{panel grid (TRUE, FALSE, or a combination of X, x, Y, y)}

\item{axis}{add x or y axes? TRUE, FALSE, "xy"}

\item{ticks}{ticks if TRUE add ticks}
}
\description{
Epurated Ggplot theme (inspired from hrbrthemes)
}
\examples{

## Not run:
library(ggplot2)
library(dplyr)

library(ggplot2)
ggplot(data = iris) +
  geom_histogram(aes(x = Sepal.Length)) +

  epi_ggtheme()

}
