% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_grid.R
\name{epi_grid}
\alias{epi_grid}
\title{Form a grid of coordinates that is inside of a map of an EpiVis data frame.}
\usage{
epi_grid(epi = NULL, pts = -1)
}
\arguments{
\item{epi}{a proper EpiVis data frame.}

\item{pts}{a numeric denoting the tentative number of points in the set. Generally points will be less than the floor of that number. Note that points = -1 if and only if kernal density is used.}
}
\value{
a set of points (a vector of coordinates)
}
\description{
Given a proper EpiVis data frame (\code{epi}) and the number of points, form a grid coordinates
in the map of the EpiVis data frame (\code{epi}).
}
\details{
Note that \code{pts} can be a positive non-integer, and the number of points in the
output grid is generally strictly less than the floor of the \code{pts},
unless the polygon is exactly rectangular.
}
