% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_plot.R
\name{epi_heatmap}
\alias{epi_heatmap}
\title{Plot a heatmap.}
\usage{
epi_heatmap(epi = NULL, method = "krig", pts = -1)
}
\arguments{
\item{epi}{a proper EpiVis data frame.}

\item{method}{a string denoting the method of interpolation. Only support bayesian kriging at the moment.}

\item{pts}{a numeric denoting the number of points to interpolate. Generally points interpolated would be less that the floor of \code{pts}.}
}
\value{
draw a heatmap on default r plot
}
\description{
High-level function for plotting a heatmap given a EpiVis dataframe.
}
\details{
It should be noted that this \code{epi_heatmap} function is faily high-leveled;
}
