% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiVis-package.R
\docType{package}
\name{EpiVis-package}
\alias{EpiVis}
\alias{EpiVis-package}
\title{EpiVis: Better Epidemic Data Visualization}
\description{
Create visualization of epidemic data by interpolating heatmap with Bayesian Kriging.
}
\author{
\strong{Maintainer}: James Zhong \email{s33zhong@uwaterloo.ca}

Authors:
\itemize{
  \item James Ni \email{mgspo1997ni@gmail.com}
}

}
\keyword{internal}
