% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{floor_week}
\alias{floor_week}
\title{Re-set date to the first day of the week the date falls in}
\usage{
floor_week(date, week_start = 1)
}
\arguments{
\item{date}{a date or something coercible by lubridate::as_date to a date}

\item{week_start}{first day of the week to re-set to. 7 = Sunday, 1 = Monday.}
}
\description{
Re-set date to the first day of the week the date falls in
}
