% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{frmt_num}
\alias{frmt_num}
\title{Format numbers is units when large}
\usage{
frmt_num(x, accuracy = 1)
}
\arguments{
\item{x}{a number to format}

\item{accuracy}{accuracy of labels, passed to \code{\link[scales:label_number_si]{scales::label_number_si}}}
}
\description{
Format numbers is units when large
}
