% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_prev_sunday}
\alias{get_prev_sunday}
\title{Get the date of the most recent Sunday}
\usage{
get_prev_sunday(date)
}
\arguments{
\item{date}{}
}
\description{
Useful for filtering data to latest full ISO week
}
