% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{label_breaks}
\alias{label_breaks}
\title{Format break labels}
\usage{
label_breaks(breaks, lab_accuracy = 1, replace_Inf = TRUE)
}
\arguments{
\item{breaks}{}

\item{lab_accuracy}{accuracy of labels, passed to \code{\link[scales:label_number_si]{scales::label_number_si}}}

\item{replace_Inf}{if \code{Inf} is your final break, replace with a + sign in the label?}
}
\description{
Format break labels
}
