% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss_vis.R
\name{miss_vis}
\alias{miss_vis}
\title{Title}
\usage{
miss_vis(
  x,
  facet = NULL,
  col_vec = c("#6a040f", "#cce3de"),
  y_axis_text_size = 8
)
}
\arguments{
\item{x}{a dataframe}

\item{facet}{a character value of variable to facet the graph}

\item{col_vec}{a vector of length 2 specifying the color for Missing and Present values respectively}

\item{y_axis_text_size}{a nuemric value for the size of the y axis text}
}
\value{
a ggplot tile graph displaying the missing/present values for all variables of the dataframe
}
\description{
Title
}
\examples{
#Use fake data from Epidemiologist R handbook

suppressMessages(c(library(dplyr), library(readr)))
linelist <- read_rds("https://github.com/appliedepi/epirhandbook/raw/master/inst/extdata/case_linelists/linelist_cleaned.rds")

plot_stacked_bar(
  linelist,
  cols = c( "fever", "vomit", "chills", "cough", "aches"),
  levels_value = c("yes"),
  keep_na = FALSE,
  use_counts = TRUE,
  flip = FALSE
)

}
