% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicurve.R
\name{plot_epicurve}
\alias{plot_epicurve}
\title{Plot incidence over time from patient level data}
\usage{
plot_epicurve(
  df,
  date_col,
  group_col = NULL,
  facet_col = NULL,
  prop_col = NULL,
  prop_numer = NULL,
  prop_denom = "non_missing",
  prop_line_colour = "black",
  prop_line_size = 0.8,
  floor_date_week = FALSE,
  label_weeks = FALSE,
  week_start = 1,
  date_breaks = "1 week",
  date_labels = waiver(),
  date_max = NULL,
  sec_date_axis = FALSE,
  facet_nrow = NULL,
  facet_ncol = NULL,
  facet_scales = "fixed",
  facet_labs = label_wrap_gen(width = 25),
  facet_lab_pos = "top",
  group_na_colour = "grey",
  title = waiver(),
  subtitle = waiver(),
  date_lab = waiver(),
  y_lab = waiver(),
  group_lab = waiver(),
  prop_lab = NULL
)
}
\arguments{
\item{df}{unaggregated dataframe with a minumum of a date column with a date or POSIX class}

\item{date_col}{date variable to plot incidence with. Must be provided.}

\item{group_col}{optional grouping variable to be applied to the fill aesthetic of columns}

\item{facet_col}{optional faceting variable to split chart into small multiples}

\item{prop_col}{optional variable to be used to plot a proportion line on top of the epicurve}

\item{prop_numer}{value(s) in the \code{prop_col} variable as a single value or vector to be used to calculate the numerator of the proportion calculation}

\item{prop_denom}{value(s) in the \code{prop_col} variable as a single value or vector to be used to calculate the denominator of the proportion calculation.
default "non_missing" will take the sum of all non-missing values in the column.}

\item{prop_line_colour}{colour of the proportion line. defaults to "red"}

\item{prop_line_size}{width of the proportion line. defaults to 0.8}

\item{floor_date_week}{should \code{date_col} dates be floored to the Monday of the ISO week they fall in? defaults to TRUE}

\item{label_weeks}{label primary date axis with week numbers? defaults to TRUE}

\item{week_start}{day of week defined as the start of the week as integer 1-7 (Monday = 1, Sunday = 7). defaults to 1 (ISO week standard)}

\item{date_breaks}{date break intervals passed to \code{\link[ggplot2:scale_date]{ggplot2::scale_x_date}}. defaults to "2 weeks"}

\item{date_labels}{\code{\link[base:strptime]{base::strptime}} date label code passed to \code{\link[ggplot2:scale_date]{ggplot2::scale_x_date}}. defaults to "\\\%V" (ISO Week)}

\item{date_max}{force a date axis max date. Useful for when a week has passed with no incidence and you want to show this
on the plot. Setting date_max to the current week will force the date axis to show this week with no incidence.}

\item{sec_date_axis}{plot a secondary date axis using default calculated ggplot2 date breaks and labels? defaults to FALSE}

\item{facet_nrow}{nrow argument passed to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap}}}

\item{facet_ncol}{ncol argument passed to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap}}}

\item{facet_labs}{facet labeller argument passed to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap}}. Defaults to \code{\link{label_wrap_gen(width = 25)}}}

\item{facet_lab_pos}{facet label position argument passed to strip.position in \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap}}. defaults to "top". Options are \code{c("top", "bottom", "left", "right")}}

\item{group_na_colour}{colour for missing values in \code{group_col}. defaults to "grey"}

\item{title}{optional title for the plot}

\item{subtitle}{optional subtitle for the plot}

\item{date_lab}{optional label for the date axis. defaults to \code{date_col} name if not provided}

\item{y_lab}{optional label for the Y axis. defaults to \code{n} if not provided}

\item{group_lab}{optional label for the group legend. defaults to \code{group_col} name if not provided}

\item{prop_lab}{label for the proportion line. There is no default so this should be provided when plotting proportion lines}
}
\value{
a ggplot object
}
\description{
Helper function to plot epidemic curves with ggplot2 with options for
grouping data, facets and proportion lines.
}
\examples{

df_ebola <- dplyr::as_tibble(outbreaks::ebola_sim_clean$linelist)

df_ebola \%>\%
  dplyr::mutate(outcome = forcats::fct_explicit_na(outcome, "Unknown")) \%>\%
  plot_epicurve(
    date_col = date_of_onset,
    group_col = outcome,
    prop_col = outcome,
    prop_numer = "Death",
    prop_denom = c("Death", "Recover"),
    floor_date_week = TRUE,
    date_breaks = "2 weeks",
    sec_date_axis = TRUE,
    date_lab = "Week of onset",
    y_lab = "Incidence",
    group_lab = "Outcome",
    prop_lab = "CFR"
  )

}
