% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss_vis.R
\name{plot_miss_vis}
\alias{plot_miss_vis}
\title{Missing data visualisation}
\usage{
plot_miss_vis(
  x,
  facet = NULL,
  col_vec = c("#6a040f", "#cce3de"),
  y_axis_text_size = 8
)
}
\arguments{
\item{x}{a dataframe}

\item{facet}{a character value of variable to facet the graph}

\item{col_vec}{a vector of length 2 specifying the color for Missing and Present values respectively}

\item{y_axis_text_size}{a nuemric value for the size of the y axis text}
}
\value{
a ggplot tile graph displaying the missing/present values for all variables of the dataframe
}
\description{
Function to generate a tile plot exploring the missing values for all observations across all variables of a dataframe.
}
\examples{
# Use simulated measles data

suppressMessages(library(dplyr))

epivis::moissala_measles |>
  filter(site \%in\% c("Moïssala Hospital", "Bouna Hospital")) |>
  plot_miss_vis(facet = "site")
}
