% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid.R
\name{plot_pyramid}
\alias{plot_pyramid}
\title{Plot Age/Sex Pyramids}
\usage{
plot_pyramid(
  df,
  age_col,
  gender_col,
  make_age_groups = TRUE,
  age_breaks = c(seq(0, 80, 10), Inf),
  age_labels = label_breaks(age_breaks),
  drop_age_levels = FALSE,
  gender_levels,
  gender_labs = NULL,
  x_lab = "Age Group",
  y_lab = "n",
  colours = c("#486090FF", "#7890A8FF"),
  show_data_labs = FALSE,
  lab_size = 4,
  lab_in_col = "white",
  lab_out_col = "grey30",
  lab_nudge_factor = 5,
  add_missing_cap = TRUE
)
}
\arguments{
\item{df}{unaggregated dataframe with a minumum of age and gender variables}

\item{add_missing_cap}{}
}
\value{
a ggplot object
}
\description{
Plot Age/Sex Pyramids
}
