% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked_bar.R
\name{plot_stacked_bar}
\alias{plot_stacked_bar}
\title{Stacked barplot}
\usage{
plot_stacked_bar(
  df,
  cols,
  levels_value,
  keep_na = TRUE,
  use_counts = TRUE,
  flip = FALSE,
  x_lab = waiver(),
  caption = TRUE
)
}
\arguments{
\item{df}{un-aggregated dataframe (linelist).}

\item{cols}{vector of character/factor variables names in \code{df} to be displayed in the barplot.}

\item{levels_value}{vector of level values to be used for the plotting.}

\item{keep_na}{logical, default = \code{TRUE.} Keep NAs in the graphs and the proportions ?}

\item{use_counts}{logical, default = \code{TRUE.} Use counts or proportion in y axis ?}

\item{flip}{logical, default = \code{FALSE}. Flip the barplot ?}

\item{x_lab}{character name for the x axis}

\item{caption}{logical, default = \code{TRUE.} Display the plot caption summarising the number of cases ?}
}
\value{
a ggplot object
}
\description{
Stacked barplot
}
\examples{

#Use fake data from Epidemiologist R handbook

suppressMessages(c(library(dplyr), library(readr)))
linelist <- read_rds("https://github.com/appliedepi/epirhandbook/raw/master/inst/extdata/case_linelists/linelist_cleaned.rds")

plot_stacked_bar(
  linelist,
  cols = c( "fever", "vomit", "chills", "cough", "aches"),
  levels_value = c("yes"),
  keep_na = FALSE,
  use_counts = TRUE,
  flip = FALSE
)

}
