% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rounder}
\alias{rounder}
\title{round number up or down to nearest 01, 100, 1000, etc}
\usage{
rounder(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{10 will round up to nearest 10, -10 down to nearest 10 etc}
}
\description{
round number up or down to nearest 01, 100, 1000, etc
}
