% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_breaks}
\alias{add_breaks}
\title{Bin numeric vector into breaks using classInt::classIntervals}
\usage{
add_breaks(x, n = 5, style = "jenks", lab_accuracy = 1,
  replace_Inf = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{n}{number of bins}

\item{style}{of class intervals. see \code{\link[classInt:classIntervals]{classInt::classIntervals}} for details}
}
\description{
Works with dplyr::mutate to add a column of breaks
}
\details{
Breaks labels are formatted with scales::label_number_si
}
