% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid.R
\name{plot_pyramid}
\alias{plot_pyramid}
\title{Plot Age/Sex Pyramids}
\usage{
plot_pyramid(
  df,
  age_col,
  gender_col,
  gender_levels,
  facet_col = NULL,
  make_age_groups = TRUE,
  age_breaks = c(seq(0, 80, 10), Inf),
  age_labels = label_breaks(age_breaks),
  drop_age_levels = FALSE,
  gender_labs = NULL,
  x_lab = waiver(),
  y_lab = waiver(),
  colours = c("#486090FF", "#7890A8FF"),
  show_data_labs = FALSE,
  lab_size = 4,
  lab_in_col = "white",
  lab_out_col = "grey30",
  lab_nudge_factor = 5,
  facet_nrow = NULL,
  facet_ncol = NULL,
  facet_scales = "fixed",
  facet_labs = label_wrap_gen(width = 25),
  facet_lab_pos = "top",
  add_missing_cap = TRUE
)
}
\arguments{
\item{df}{un-aggregated dataframe with a minimum of age and gender variables.}

\item{age_col}{age variable name in \code{df}. Can be either a numeric vecotr of ages
or a character/factor vector of age groups.}

\item{gender_col}{gender variable name in \code{df} with levels indicating male or female.}

\item{gender_levels}{length 2 character vector with male and female level in \code{gender_col}, respectively.}

\item{facet_col}{optional faceting variable name to split chart into small multiples.}

\item{make_age_groups}{set to TRUE (default) if \code{age_col} is numeric and needs to be binned into groups.}

\item{age_breaks}{breaks to be used for binning a numerical \code{age_col}.}

\item{age_labels}{break labels to accompany \code{age_breaks}. Defaults to \code{epivis::label_breaks(age_breaks)}.}

\item{drop_age_levels}{should age groups with no observations be removed from the chart? Defaults to FALSE.}

\item{gender_labs}{optional labels for \code{gender_levels}}

\item{x_lab}{optional label for the X axis.}

\item{y_lab}{optional label for the Y axis.}

\item{colours}{length 2 character vector of colours used for male and female, respectively.}

\item{show_data_labs}{show data labels on chart? Defaults to FALSE.}

\item{lab_size}{data labels size.}

\item{lab_in_col}{data label colour when placed inside a bar.}

\item{lab_out_col}{data label colour when placed outside a bar.}

\item{lab_nudge_factor}{threshold for moving a data label outside a bar. Defaults to 5.
Increasing the number increases the distance from the max value required to move a label outside the bar.}

\item{facet_nrow}{nrow argument passed to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap}}}

\item{facet_ncol}{ncol argument passed to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap}}}

\item{facet_scales}{facet scales argument passed to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap}}.
Should scales be fixed ("fixed", the default), free ("free"), or free in one dimension ("free_x", "free_y")?}

\item{facet_labs}{facet labeller argument passed to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap}}. Defaults to \code{\link{label_wrap_gen(width = 25)}}}

\item{facet_lab_pos}{facet label position argument passed to strip.position in \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap}}.
Defaults to "top". Options are \code{c("top", "bottom", "left", "right")}.}

\item{add_missing_cap}{show missing data counts for \code{age_col} and \code{gender_col}? Defaults to TRUE.}
}
\value{
a ggplot object
}
\description{
Plot Age/Sex Pyramids
}
\examples{

df_flu <- outbreaks::fluH7N9_china_2013

plot_pyramid(
  df = df_flu,
  age_col = age,
  gender_col = gender,
  gender_levels = c("m", "f")
)

}
